2.0.2
- behobenes Problem mit Formatwandlung int in BigDecimal für Hibernate / behoben in Angebot, Bestellung, Rechnung und Lieferschein
- Lieferschein implementiert und zur Nutzung freigegeben
- SUPERUSER hat jetzt alle Tabs (inkl. Jahresergebnis)

+++ src/org/andy/code/dataExport/ExcelAngebot.java
@@ -91,7 +91,7 @@ public class ExcelAngebot implements Identified {
 			//#######################################################################
 			// Zellen in Tabelle Enummerieren
 			//#######################################################################
-			for(int i = 0; i < angebot.getAnzPos().intValue(); i++ ) { //Angebotspositionen B, C, D, F Zeile 17-28
+			for(int i = 0; i < angebot.getAnzPos(); i++ ) { //Angebotspositionen B, C, D, F Zeile 17-28
 				int j = i + START_ROW_OFFSET;
 				anPos[i] = ws.getRow(j).getCell(COLUMN_A); //Position
 				anText[i] = ws.getRow(j).getCell(COLUMN_B); //Text
@@ -109,7 +109,7 @@ public class ExcelAngebot implements Identified {
 			ExportHelper.replaceCellValue(wb, ws, "{anRef}", angebot.getRef());
 			ExportHelper.replaceCellValue(wb, ws, "{anDuty}", kunde.getPronomen() + " " + kunde.getPerson());
 			
-			for(int i = 0; i < angebot.getAnzPos().intValue(); i++ ) {
+			for(int i = 0; i < angebot.getAnzPos(); i++ ) {
 				anPos[i].setCellValue(String.valueOf(i + 1));
 				try {
 					String art = (String) Angebot.class.getMethod("getArt" + String.format("%02d", i + 1)).invoke(angebot);
